command: Fix panic for nil credentials source
If loading credentials fails, the credentials source can be nil. Calling
any of its methods would then result in a panic. This commit adds nil
receiver checks to each of the public methods to cope with this case.

The user-facing bug that triggered this work is that running init when
the HOME environment variable is unset would result in a panic on macOS.